Added early stopping
Added early stopping instead of the old way of running for a set amount of epochs, this should ensure a more complete fitment. Also moved loss and data, as well as the early stopping to a seperate helper folder for better readability and implementability in the models. Also fixed some minor mistakes and removed unused code