Archetypical analisys now works!!!
The dimensions in last version where wrong.
This version fixes it but switches the order of multiplication of the matrices. The dimension of the matrices are now agreeing with the "Archetypal analysis for machine learning and data mining" paper but reversed so it works with how our data is structured NxM, samples in the rows and features in the columns.